remove unneeded components
